bert4rec/requirements.txt
tqdm==4.65.0
torch==2.0.1
python-box==7.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.3
seaborn==0.12.2
numpy==1.24.3
pandas==2.0.2